S15.d.1: extract up.ts --upgrade fast-path (1849 → 1660 LOC)
First sub-slice of the up.ts multi-PR sub-train. Extract the
self-contained `--upgrade` fast-path (cached-context Helm rerun +
sandbox fed-cred sync) to cli/src/commands/up/fast_upgrade.ts.

Caller invokes via dynamic import:
  if (options.upgrade) {
    const { runFastUpgrade } = await import("./up/fast_upgrade.js");
    await runFastUpgrade(options);
    return;
  }

Verification:
- up.ts: 1849 → 1660 LOC (further reduction in d.2-d.4)
- tsc --noEmit / lint (27 warnings, baseline) / build: clean
- npm test: 454 passed; 2 skipped pre-existing

Subsequent sub-slices d.2-d.4 will continue toward the §15 800 cap:
- d.2: preflight extraction
- d.3: Helm + AgentMesh deploy phases
- d.4: sandbox bring-up + summary

No behavioral change: `if (options.upgrade)` body moved verbatim.

Audit: docs/security-audits/2026-04-29-phase2-hotspot-up-cli.md
